17 exception handler inventory service

+++ inventory/repositories.py
@@ -23,11 +23,19 @@ class AuthorRepository:
     def get_books_by_id(self, id_author: int):
         stmt = select(Book).where(Book.author_id == id_author)
         result = self.session.execute(stmt).scalars().all()
+        if result == None:
+            raise NotFoundException(id_author, Author.__name__)
         return result
         
-    def delete_by_id(self, id_author: int) -> None:
-        self.session.delete(Author, id_author)
+    def delete_by_id(self, id_author: int) -> Author:
+        stmt = select(Author).where(Author.id == id_author)
+        result = self.session.execute(stmt).scalars().first()
+        if result == None:
+            raise NotFoundException(id_author, Author.__name__)
+        
+        self.session.delete(result) 
         self.session.commit()
+        return result
 
     def check_author(self, author: AuthorCreate):
         stmt = select(Author).where(Author.name == author.name) 
@@ -42,9 +50,9 @@ class AuthorRepository:
             self.session.refresh(new_author)
 
             author_resp = AuthorResponse(
-            id = new_author.author.id,
-            name = new_author.author.name,
-            birthday = new_author.author.birthday
+            id = new_author.id,
+            name = new_author.name,
+            birthday = new_author.birthday
         )
 
             return author_resp
@@ -59,6 +67,8 @@ class AuthorRepository:
     def update(self, id: int, new_author: Author):
         stmt = select(Author).where(Author.id == id)
         result :Author = self.session.execute(stmt).scalars().first()
+        if result == None:
+            raise NotFoundException(id, Author.__name__)
         for key,value in dict(new_author).items():
             if key != "id":
                 if hasattr(result, key):
@@ -100,9 +110,11 @@ class BooksRepository:
     def delete_by_isbn(self, isbn: str):
         stmt = select(Book).where(Book.isbn == isbn) 
         result = self.session.execute(stmt).scalars().first()
+        if result == None:
+            raise NotFoundException(isbn, Book.__name__)
         self.session.delete(result)
         self.session.commit()
-        return result   
+        return f" Book {result} is deleted"   
     
     def mapping_book(self, book: Book):
         new_book = Book(
@@ -118,6 +130,7 @@ class BooksRepository:
             stock = book.stock
 
         )
+
         self.session.add(new_book)
         self.session.commit()
         self.session.refresh(new_book)
@@ -133,14 +146,14 @@ class BooksRepository:
             release = new_book.release,
             genre= GenreResponse(
                 id = new_book.genre.id,
-                genre = new_book.genre.genre
+                name = new_book.genre.name
             ),
             description = new_book.description,
             price = new_book.price,
             age_recommendation = new_book.age_recommendation,
             publisher = PublisherResponse(
                 id = new_book.publisher.id,
-                publisher = new_book.publisher.publisher
+                name = new_book.publisher.name
             ),
             stock = new_book.stock
         )
@@ -149,6 +162,8 @@ class BooksRepository:
     def update(self, isbn: str, new_book: Book):
         stmt = select(Book).options(subqueryload(Book.author)).where(Book.isbn == isbn)
         result :Book = self.session.execute(stmt).scalars().first()
+        if result == None:
+            raise NotFoundException(isbn, Book.__name__)
         for key,value in dict(new_book).items():
             if key != "isbn":
                 if hasattr(result, key):
@@ -174,19 +189,31 @@ class GenreRepository:
         return genres
     
     def get_by_id(self, id_genre: int):
-        return self.session.get(Genre, id_genre)
+        stmt = select(Genre).where(Genre.id == id_genre)
+        result = self.session.execute(stmt).scalars().all()
+        if result == None:
+            raise NotFoundException(id_genre, Genre.__name__)
+        return result
     
     def get_books_by_id(self, id_genre: int):
         stmt = select(Book).where(Book.genre_id == id_genre)
         result = self.session.execute(stmt).scalars().all()
+        if result == None:
+            raise NotFoundException(id_genre, Genre.__name__)
         return result
         
-    def delete_by_id(self, id_genre: int) -> None:
-        self.session.delete(Genre, id_genre)
+    def delete_by_id(self, id_genre: int):
+        stmt = select(Genre).where(Genre.id == id_genre)
+        result = self.session.execute(stmt).scalars().first()
+        if result == None:
+            raise NotFoundException(id_genre, Genre.__name__)
+        
+        self.session.delete(result) 
         self.session.commit()
+        return result
 
     def check_genre(self, genre: GenreCreate):
-        stmt = select(Genre).where(Genre.genre == genre) 
+        stmt = select(Genre).where(Genre.name == genre.name) 
         result = self.session.exec(stmt).scalars().first()
         if not result:
             new_genre = Genre(
@@ -212,6 +239,8 @@ class GenreRepository:
     def update(self, id: int, new_genre: Genre):
         stmt = select(Genre).where(Genre.id == id)
         result :Genre = self.session.execute(stmt).scalars().first()
+        if result == None:
+            raise NotFoundException(id_genre, Genre.__name__)
         for key,value in dict(new_genre).items():
             if key != "id":
                 if hasattr(result, key):
@@ -233,31 +262,39 @@ class PublisherRepository:
         self.session = session
 
     def get_all(self):
-        #with Session(self.engine) as s:
         publishers = self.session.query(Publisher).all()
-            # stmt = select(Book)
-            # result = s.execute(stmt)
-            # books = result.all()
         return publishers
     
     def get_by_id(self, id_publisher: int):
-        return self.session.get(Publisher, id_publisher)
+        stmt = select(Publisher).where(Publisher.id == id_publisher)
+        result = self.session.execute(stmt).scalars().all()
+        if result == None:
+            raise NotFoundException(id_publisher, Publisher.__name__)
+        return result
     
     def get_books_by_id(self, id_publisher: int):
         stmt = select(Book).where(Book.publisher_id == id_publisher)
         result = self.session.execute(stmt).scalars().all()
+        if result == None:
+            raise NotFoundException(id_publisher, Publisher.__name__)
         return result
         
     def delete_by_id(self, id_publisher: int) -> None:
-        self.session.delete(Publisher, id_publisher)
+        stmt = select(Publisher).where(Publisher.id == id_publisher)
+        result = self.session.execute(stmt).scalars().first()
+        if result == None:
+            raise NotFoundException(id_publisher, Publisher.__name__)
+        
+        self.session.delete(result) 
         self.session.commit()
+        return result
     
     def check_publisher(self, publisher: PublisherCreate):
-        stmt = select(Publisher).where(Publisher.genre == publisher) 
+        stmt = select(Publisher).where(Publisher.name == publisher.name) 
         result = self.session.exec(stmt).scalars().first()
         if not result:
             new_publisher = Publisher(
-                publisher = publisher.publisher,
+                name = publisher.name,
             )
             self.session.add(new_publisher)
             self.session.commit()
@@ -295,6 +332,8 @@ class PublisherRepository:
     def update(self, id: int, new_publisher: Publisher):
         stmt = select(Publisher).where(Publisher.id == id)
         result :Publisher = self.session.execute(stmt).scalars().first()
+        if result == None:
+            raise NotFoundException(id_publisher, Publisher.__name__)
         for key,value in dict(new_publisher).items():
             if key != "id":
                 if hasattr(result, key):
